eLibrary/src/backend/requirement.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
wikipedia
pymupdf
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
py
transformer
torch
os
pdfplumber
